refactored plugin code (split code into multiple files; added dataStore service)
