+++ src/plugins/timezones/components/TimezoneModal.tsx
@@ -4,23 +4,17 @@
  * SPDX-License-Identifier: GPL-3.0-or-later
  */
 
-import * as DataStore from "@api/DataStore";
 import { classNameFactory } from "@api/Styles";
 import { Margins } from "@utils/margins";
 import { ModalCloseButton, ModalContent, ModalFooter, ModalHeader, ModalProps, ModalRoot } from "@utils/modal";
 import { Button, Forms, SearchableSelect, useMemo, useState } from "@webpack/common";
 
-import { DATASTORE_KEY, timezones } from ".";
-
-export async function setUserTimezone(userId: string, timezone: string | null) {
-    timezones[userId] = timezone;
-    await DataStore.set(DATASTORE_KEY, timezones);
-}
+import { timezonesStoreService } from "../index";
 
 const cl = classNameFactory("vc-timezone-");
 
 export function SetTimezoneModal({ userId, modalProps }: { userId: string, modalProps: ModalProps; }) {
-    const [currentValue, setCurrentValue] = useState<string | null>(timezones[userId] ?? null);
+    const [currentValue, setCurrentValue] = useState<string | null>(timezonesStoreService.getUserTimezone(userId) ?? null);
 
     const options = useMemo(() => {
         return Intl.supportedValuesOf("timeZone").map(timezone => {
@@ -62,7 +56,7 @@ export function SetTimezoneModal({ userId, modalProps }: { userId: string, modal
                 <Button
                     color={Button.Colors.RED}
                     onClick={async () => {
-                        await setUserTimezone(userId, null);
+                        await timezonesStoreService.setUserTimezone(userId, null);
                         modalProps.onClose();
                     }}
                 >
@@ -72,7 +66,7 @@ export function SetTimezoneModal({ userId, modalProps }: { userId: string, modal
                     color={Button.Colors.BRAND}
                     disabled={currentValue === null}
                     onClick={async () => {
-                        await setUserTimezone(userId, currentValue!);
+                        await timezonesStoreService.setUserTimezone(userId, currentValue!);
                         modalProps.onClose();
                     }}
                 >
