+++ src/plugins/timezones/types.ts
@@ -1,0 +1,11 @@
+/*
+ * Vencord, a Discord client mod
+ * Copyright (c) 2024 Vendicated and contributors
+ * SPDX-License-Identifier: GPL-3.0-or-later
+ */
+
+export interface LocalUserTimestamp {
+    userId: string;
+    timestamp?: string;
+    type: "message" | "profile";
+}
