+++ src/plugins/timezones/utils.ts
@@ -1,0 +1,19 @@
+/*
+ * Vencord, a Discord client mod
+ * Copyright (c) 2024 Vendicated and contributors
+ * SPDX-License-Identifier: GPL-3.0-or-later
+ */
+
+import { i18n } from "@webpack/common";
+
+import { settings } from "./settings";
+
+export function getTime(timezone: string, timestamp: string | number, props: Intl.DateTimeFormatOptions = {}) {
+    const date = new Date(timestamp);
+    const formatter = new Intl.DateTimeFormat(i18n?.getLocale?.() ?? "en-US", {
+        hour12: !settings.store["24h Time"],
+        timeZone: timezone,
+        ...props
+    });
+    return formatter.format(date);
+}
